+++ deploy/2_registration.migration.ts
@@ -11,7 +11,7 @@ import {
   deployCECDSADispatcher,
 } from "./helpers";
 
-import { Registration2Mock__factory, StateKeeperMock__factory } from "@ethers-v6";
+import { Registration2Mock__factory, RegistrationSimple__factory, StateKeeperMock__factory } from "@ethers-v6";
 
 import { getConfig } from "./config/config";
 
@@ -21,6 +21,14 @@ export = async (deployer: Deployer) => {
 
   const registration = await deployProxy(deployer, Registration2Mock__factory, "Registration2");
   await registration.__Registration_init(config.tssSigner, config.chainName, await stateKeeper.getAddress());
+
+  const registrationSimple = await deployProxy(deployer, RegistrationSimple__factory, "RegistrationSimple");
+  await registrationSimple.__RegistrationSimple_init(
+    config.tssSigner,
+    config.chainName,
+    await stateKeeper.getAddress(),
+    config.simpleRegistrationSigners,
+  );
 
   await deployPVerifiers(deployer);
 
@@ -48,5 +56,8 @@ export = async (deployer: Deployer) => {
   await deployPNOAADispatcher(deployer);
   await deployPECDSASHA12704Dispatcher(deployer);
 
-  Reporter.reportContracts(["Registration2", `${await registration.getAddress()}`]);
+  Reporter.reportContracts(
+    ["Registration2", `${await registration.getAddress()}`],
+    ["RegistrationSimple", `${await registrationSimple.getAddress()}`],
+  );
 };